+++ src/EventListener/ContentSecurityPolicyListener.php
@@ -23,8 +23,6 @@ use Symfony\Component\HttpKernel\KernelEvents;
 
 final class ContentSecurityPolicyListener extends AbstractContentTypeRestrictableListener
 {
-    use KernelEventForwardCompatibilityTrait;
-
     private DirectiveSet $report;
     private DirectiveSet $enforce;
     private bool $compatHeaders;
@@ -68,7 +66,7 @@ final class ContentSecurityPolicyListener extends AbstractContentTypeRestrictabl
 
     public function onKernelRequest(RequestEvent $e): void
     {
-        if (!$this->isMainRequest($e)) {
+        if (!$e->isMainRequest()) {
             return;
         }
 
@@ -135,7 +133,7 @@ final class ContentSecurityPolicyListener extends AbstractContentTypeRestrictabl
 
     public function onKernelResponse(ResponseEvent $e): void
     {
-        if (!$this->isMainRequest($e)) {
+        if (!$e->isMainRequest()) {
             return;
         }